model/sd/requirements.txt
diffusers
python-dotenv
torch
websockets==12